+++ packages/design-system/CHANGELOG.md
@@ -17,6 +17,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 - **FormSelect:** ability to forward react ref to input.
 - **FormTextArea:** ability to forward react ref to input.
 - **FormFeedback:** remove `span: 9` from styles.
+- **DropdownButtonItem:** extend Props Interface to accept `as` prop.
 
 # [1.1.0](https://github.com/IQSS/dataverse-frontend/compare/@iqss/dataverse-design-system@1.0.1...@iqss/dataverse-design-system@1.1.0) (2024-03-12)